feat: Implement custom scroll physics with QrioScrollPhysics and replace existing physics in History widget
