+++ lib/src/widgets/history.dart
@@ -1,7 +1,6 @@
 import 'dart:convert';
 import 'dart:math' as math;
 
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
@@ -9,6 +8,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:qrio/src/app.dart';
 import 'package:qrio/src/constants.dart';
 import 'package:qrio/src/models/history_model.dart';
+import 'package:qrio/src/physics/qrio_scroll_physics.dart';
 import 'package:qrio/src/utils.dart';
 import 'package:qrio/src/widgets/data_bottom_sheet.dart';
 import 'package:qrio/src/widgets/history_menu_sheet.dart';
@@ -282,7 +282,7 @@ class History extends ConsumerWidget {
                 height: scrollContentHeight,
                 child: ListView.builder(
                   controller: controller,
-                  physics: const CustomBouncingScrollPhysics(),
+                  physics: const QrioScrollPhysics(),
                   padding: const EdgeInsets.fromLTRB(0, 8, 0, 80),
                   itemCount: hisLen,
                   itemBuilder: (context, i) {
@@ -474,98 +474,4 @@ class History extends ConsumerWidget {
       },
     );
   }
-}
-
-class CustomBouncingScrollPhysics extends ScrollPhysics {
-  const CustomBouncingScrollPhysics({
-    this.decelerationRate = ScrollDecelerationRate.normal,
-    super.parent,
-  });
-
-  final ScrollDecelerationRate decelerationRate;
-
-  @override
-  CustomBouncingScrollPhysics applyTo(ScrollPhysics? ancestor) {
-    return CustomBouncingScrollPhysics(
-        parent: buildParent(ancestor), decelerationRate: decelerationRate);
-  }
-
-  double frictionFactor(double overscrollFraction) {
-    return 0.01 * math.pow(1 - overscrollFraction, 2);
-  }
-
-  @override
-  double applyPhysicsToUserOffset(ScrollMetrics position, double offset) {
-    assert(offset != 0.0);
-    assert(position.minScrollExtent <= position.maxScrollExtent);
-
-    if (!position.outOfRange) {
-      return offset;
-    }
-
-    final double overscrollPastStart =
-        math.max(position.minScrollExtent - position.pixels, 0.0);
-    final double overscrollPastEnd =
-        math.max(position.pixels - position.maxScrollExtent, 0.0);
-    final double overscrollPast =
-        math.max(overscrollPastStart, overscrollPastEnd);
-    final bool easing = (overscrollPastStart > 0.0 && offset < 0.0) ||
-        (overscrollPastEnd > 0.0 && offset > 0.0);
-
-    final double friction = easing
-        ? frictionFactor(
-            (overscrollPast - offset.abs()) / position.viewportDimension)
-        : frictionFactor(overscrollPast / position.viewportDimension);
-    final double direction = offset.sign;
-
-    return direction * _applyFriction(overscrollPast, offset.abs(), friction);
-  }
-
-  static double _applyFriction(
-      double extentOutside, double absDelta, double gamma) {
-    assert(absDelta > 0);
-    double total = 0.0;
-    if (extentOutside > 0) {
-      final double deltaToLimit = extentOutside / gamma;
-      if (absDelta < deltaToLimit) {
-        return absDelta * gamma;
-      }
-      total += extentOutside;
-      absDelta -= deltaToLimit;
-    }
-    return total + absDelta;
-  }
-
-  @override
-  double applyBoundaryConditions(ScrollMetrics position, double value) => 0.0;
-
-  @override
-  Simulation? createBallisticSimulation(
-      ScrollMetrics position, double velocity) {
-    final Tolerance tolerance = toleranceFor(position);
-    if (velocity.abs() >= tolerance.velocity || position.outOfRange) {
-      return BouncingScrollSimulation(
-          spring: spring,
-          position: position.pixels,
-          velocity: velocity,
-          leadingExtent: position.minScrollExtent,
-          trailingExtent: position.maxScrollExtent,
-          tolerance: tolerance,
-          constantDeceleration: 0);
-    }
-    return null;
-  }
-
-  @override
-  double get minFlingVelocity => kMinFlingVelocity * 2.0;
-
-  @override
-  double carriedMomentum(double existingVelocity) {
-    return existingVelocity.sign *
-        math.min(0.000816 * math.pow(existingVelocity.abs(), 1.967).toDouble(),
-            40000.0);
-  }
-
-  @override
-  double get dragStartDistanceMotionThreshold => 3.5;
 }
